Added a WBP for a canvas and blueprint. font and padding changes. Showed WBP on the character BP.
Adding a WBP for the message box with a border drawn as a box with 0.2 margin and 20 padding. Added a text box as a child to that border.
Adding a WBP for the MainHUD with a canvas (should use the least amount of canvases possible for optimization) with the WBP for the message box as a child. Changed various setting to that like the anchor point, y position, and offset right and left. Made some changes to the BP_Red to show to canvas on eventBeginPlay.
Changed the font and applied auto wrap on the font as well.
Set visibility of the canvas to hidden