Adding pickup functionality
Added a Actor BP for picking up things. Added the chest and changed some sprite settings for it.
Added a box collision to BP_Red and the BP_Pickup.
Added IA_Interact which is for checking if the player collider is colliding with a interactable object.
Added some variables for item name and item id and a function to destroy the bp_pickup actor and give me those variables back from the BP_Red